+++ engines/xeen/map.cpp
@@ -479,7 +479,7 @@ void MonsterObjectData::synchronize(XeenSerializer &s, MonsterData &monsterData)
 			obj._direction = mobStruct._direction;
 			obj._frame = 100;
 
-			if (obj._id < (int)_objectSprites.size()) {
+			if (obj._id >= 0 && obj._id < (int)_objectSprites.size()) {
 				obj._spriteId = _objectSprites[obj._id]._spriteId;
 				obj._sprites = &_objectSprites[obj._id]._sprites;
 			}
@@ -509,6 +509,12 @@ void MonsterObjectData::synchronize(XeenSerializer &s, MonsterData &monsterData)
 
 			if (mon._id < (int)_monsterSprites.size()) {
 				mon._spriteId = _monsterSprites[mon._id]._spriteId;
+
+				// WORKAROUND: For Clouds-only mine Alpha, the monster list is invalid
+				if (_vm->getGameID() == GType_Clouds && mon._spriteId == 105
+					&& monsterData.size() == 91)
+					return;
+
 				mon._sprites = &_monsterSprites[mon._id]._sprites;
 				mon._attackSprites = &_monsterSprites[mon._id]._attackSprites;
 				mon._monsterData = &monsterData[mon._spriteId];