+++ NEWS.md
@@ -146,6 +146,9 @@ For a more comprehensive changelog of the latest experimental code, see:
    - Adjust cursor to behave closer to the original.
    - Adjust item quantity slider to behave closer to the original.
 
+ Xeen:
+   - Several crash fixes for Clouds of Xeen
+
  3DS port:
    - Update relocation parser to support PREL31 that are emitted by new compiler.
    - Do more autoconfiguration in ./configure.